Update Ultroid v0.8
---------

+++ plugins/beautify.py
@@ -10,11 +10,12 @@ from . import get_help
 __doc__ = get_help("help_beautify")
 
 
+import os
 import random
 
 from telethon.utils import get_display_name
 
-from . import Carbon, eor, get_string, inline_mention, os, ultroid_cmd
+from . import Carbon, ultroid_cmd, get_string, inline_mention
 
 _colorspath = "resources/colorlist.txt"
 
@@ -28,7 +29,7 @@ else:
 @ultroid_cmd(
     pattern="(rc|c)arbon",
 )
-async def crbn(event):
+async def cr_bn(event):
     xxxx = await event.eor(get_string("com_1"))
     te = event.pattern_match.group(1)
     col = random.choice(all_col) if te[0] == "r" else "White"
@@ -45,8 +46,11 @@ async def crbn(event):
         try:
             code = event.text.split(" ", maxsplit=1)[1]
         except IndexError:
-            return await eor(xxxx, get_string("carbon_2"))
+            return await xxxx.eor(get_string("carbon_2"))
     xx = await Carbon(code=code, file_name="ultroid_carbon", backgroundColor=col)
+    if isinstance(xx, dict):
+        await xxxx.edit(f"`{xx}`")
+        return
     await xxxx.delete()
     await event.reply(
         f"Carbonised by {inline_mention(event.sender)}",
@@ -77,7 +81,7 @@ async def crbn(event):
             code = match[1]
             match = match[0]
         except IndexError:
-            return await eor(msg, get_string("carbon_2"))
+            return await msg.eor(get_string("carbon_2"))
     xx = await Carbon(code=code, backgroundColor=match)
     await msg.delete()
     await event.reply(
@@ -125,7 +129,7 @@ async def pass_on(ult):
         theme = random.choice(RaySoTheme)
     if ult.is_reply:
         msg = await ult.get_reply_message()
-        text = msg.text
+        text = msg.message
         title = get_display_name(msg.sender)
     await ult.reply(
         file=await Carbon(text, rayso=True, title=title, theme=theme, darkMode=dark)
